#77-migrating-the-project-to-clean-architecture - Restructured the code
