+++ lib/core/data/models/queue_item/queue_item.dart
