+++ lib/features/generative_ai/presentation/provider/gemini_provider.dart
