+++ lib/features/routing_features/shell_route/shell_child_one/shell_child_one.dart
@@ -2,7 +2,7 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
-import 'package:sample_latest/provider/route_provider.dart';
+import 'package:sample_latest/core/presentation/provider/route_provider.dart';
 
 class ShellChildOneChildOne extends StatelessWidget {
   const ShellChildOneChildOne({Key? key}) : super(key: key);
