+++ lib/features/routing_features/shell_route/shell_child_one/shell_child_three.dart
@@ -1,7 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:sample_latest/provider/route_provider.dart';
+import 'package:sample_latest/core/presentation/provider/route_provider.dart';
 
 class ShellChildOneChildThree extends StatelessWidget {
   const ShellChildOneChildThree({Key? key}) : super(key: key);
