+++ lib/features/routing_features/shell_route/shell_routing.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
-import 'package:sample_latest/provider/route_provider.dart';
+import 'package:sample_latest/core/presentation/provider/route_provider.dart';
 import 'package:sample_latest/core/widgets/custom_app_bar.dart';
 
 class ShellRouting extends StatelessWidget {
