+++ lib/main.dart
@@ -11,13 +11,13 @@ import 'package:loader_overlay/loader_overlay.dart';
 import 'package:provider/provider.dart';
 import 'package:sample_latest/features/daily_tracker/presentation/bloc/daily_status_tracker_bloc.dart';
 import 'package:sample_latest/features/schools/presentation/bloc/school_bloc.dart';
-import 'package:sample_latest/provider/gemini_provider.dart';
+import 'package:sample_latest/features/generative_ai/presentation/provider/gemini_provider.dart';
 import 'package:sample_latest/core/data/db/db_configuration.dart';
 import 'package:sample_latest/features/daily_tracker/data/repository/daily_tracker_repository.dart';
 import 'package:sample_latest/features/schools/data/repository/school_repository.dart';
 import 'package:sample_latest/global_variables.dart';
 import 'package:sample_latest/latest_3.0.dart';
-import 'package:sample_latest/provider/common_provider.dart';
+import 'package:sample_latest/core/presentation/provider/common_provider.dart';
 import 'package:sample_latest/core/routing.dart';
 import 'package:sample_latest/core/theme.dart';
 import 'package:sample_latest/features/push_notifcations/push_notification_service.dart';
