+++ test/widget_testing/school/view_student_details_test.dart
@@ -8,7 +8,7 @@ import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
 import 'package:sample_latest/features/schools/presentation/bloc/school_bloc.dart';
 import 'package:sample_latest/core/environment/environment.dart';
-import 'package:sample_latest/provider/common_provider.dart';
+import 'package:sample_latest/core/presentation/provider/common_provider.dart';
 import 'package:sample_latest/features/schools/presentation/screens/student.dart';
 import 'package:sample_latest/utils/device_configurations.dart';
